feat: refactor to a Sidebar component
